change interface into abstract class
